+++ apps/community/CalendarSync/Calendar.php
@@ -6,7 +6,7 @@ use HubletoApp\Community\CalendarSync\Models\Source;
 
 class Calendar extends \HubletoMain\Core\Calendar {
 
-  public function loadEvents(): array
+  public function loadEvents(string $dateStart, string $dateEnd): array
   {
     $formattedEvents = [];
     $mSources = new Source($this->main);